Añadido comentario del ejercicio 1 con las conversiones y iniciado ejercicio 6 de algoritmos secuenciales

+++ src/main/java/algoritmossecunciales/Ej01.java
@@ -30,13 +30,13 @@ public class Ej01 {
 
         //solicitos los datos 
         System.out.println("Indique los grados que quiere transformar ");
-        gradosPedidos = entrada.nextInt();
+        gradosPedidos = entrada.nextDouble();
         //proceso
         resultado = (gradosPedidos * CONVERSOR_FAHRENHEIT) + CONVERSOR_FAHRENHEIT2;
 
         // muestro los datos
         System.out.println(gradosPedidos + " son = " + resultado + " Fahrenheit");
-
+           
     }
 
 }
